:recycle: Refactor functions into TypeScript
