+++ functions/.gitignore
@@ -1,3 +1,12 @@
+## Compiled JavaScript files
+**/*.js
+**/*.js.map
+
+# Typescript v1 declaration files
+typings/
+
 node_modules/
+
+# Custom files
 .runtimeconfig.json
 sk.json
