+++ functions/tsconfig.json
@@ -1,0 +1,17 @@
+{
+	"compilerOptions": {
+		"module": "commonjs",
+		"noImplicitReturns": true,
+		"noUnusedLocals": true,
+		"outDir": "lib",
+		"sourceMap": true,
+		"strict": true,
+		"target": "es2017",
+		"types": []
+	},
+	"compileOnSave": true,
+	"include": [
+		"./src"
+	],
+	"exclude": ["../node_modules"]
+}
